# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/latest/hr-users/{username}:
    get:
      tags:
      - hr-controller
      operationId: getUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
    put:
      tags:
      - hr-controller
      operationId: updateUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/latest/user-sessions:
    get:
      tags:
      - user-session-controller
      operationId: getUserSession
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSessionDto"
    post:
      tags:
      - user-session-controller
      operationId: createUserSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSessionRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSessionDto"
  /api/latest/user-sessions/generate-long-term-goal:
    post:
      tags:
      - user-session-controller
      operationId: generateLongTermGoal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionSteps"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
  /api/latest/user-sessions/generate-action-steps:
    post:
      tags:
      - user-session-controller
      operationId: generateActionStep
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionSteps"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
  /api/latest/user-sessions/feedback:
    post:
      tags:
      - user-session-controller
      operationId: setFeedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeedbackDto"
...
  /api/latest/user-info/locale:
    post:
      tags:
      - user-info-controller
      operationId: setLocale
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLocaleRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserInfoDto"
...
  /api/latest/user-sessions/generate-recommended-growth:
    get:
      tags:
      - user-session-controller
      operationId: generateRecommendedGrowth
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendedGrowth"
...
components:
  schemas:
    ErrorCodeFieldDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
    ErrorDto:
      type: object
      properties:
        errorCode:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorCodeFieldDto"
        errorMessage:
          type: string
...
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        username:
          type: string
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        timeZone:
          type: string
          minLength: 1
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          minItems: 1
          uniqueItems: true
        locale:
          type: string
          pattern: "[a-z]{2}"
        manager:
          type: string
        isManager:
          type: boolean
        position:
          type: string
        aspiredCompetency:
          type: string
      required:
      - authorities
      - firstName
      - lastName
      - status
      - timeZone
    UserDto:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        timeZone:
          type: string
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        position:
          type: string
        manager:
          type: string
        isManager:
          type: boolean
        aspiredCompetency:
          type: string
...
    NewActionStepDto:
      type: object
      properties:
        label:
          type: string
        date:
          type: string
          format: date
    UserSessionRequest:
      type: object
      properties:
        areaOfDevelopment:
          type: array
          items:
            type: string
          minItems: 1
        longTermGoal:
          type: string
          maxLength: 1000
          minLength: 1
        motivation:
          type: string
          maxLength: 2000
          minLength: 1
        actionSteps:
          type: array
          items:
            $ref: "#/components/schemas/NewActionStepDto"
      required:
      - areaOfDevelopment
    ActionStepDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        date:
          type: string
          format: date
        checked:
          type: boolean
    UserSessionDto:
      type: object
      properties:
        areaOfDevelopment:
          type: array
          items:
            type: string
        longTermGoal:
          type: string
        motivation:
          type: string
        actionSteps:
          type: array
          items:
            $ref: "#/components/schemas/ActionStepDto"
        lastReflection:
          type: string
    ActionSteps:
      type: object
      properties:
        areaOfDevelopment:
          type: string
        longTermGoal:
          type: string
    FeedbackDto:
      type: object
      properties:
        username:
          type: string
        sessionId:
          type: integer
          format: int64
        answers:
          type: object
          additionalProperties:
            type: integer
            format: int32
        feedback:
          type: string
      required:
      - answers
...
    UserInfoDto:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        userRoles:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        timeZone:
          type: string
        strengths:
          type: array
          items:
            type: string
        values:
          type: array
          items:
            type: string
        areaOfDevelopment:
          type: array
          items:
            type: string
        coach:
          type: string
        locale:
          type: string
        allowedCoachRates:
          type: array
          items:
            type: string
          uniqueItems: true
        companyId:
          type: integer
          format: int64
...
    SetLocaleRequestDto:
      type: object
      properties:
        locale:
          type: string
          pattern: "[a-z]{2}"
...
    RecommendedGrowth:
      type: object
      properties:
        area:
          type: string
        recommendation:
          type: string
...